% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-mhc.R
\name{anchorPositions}
\alias{anchorPositions}
\title{Determine Anchor Positions}
\usage{
anchorPositions(mhc = "HLA-A-02:01", l = 9, k = 2)
}
\arguments{
\item{mhc}{the mhc molecules.}

\item{l}{the peptide length.}

\item{k}{how many anchor positions to return.}
}
\description{
Uses the information in the SMM matrices to predict anchor positions for a given MHC
molecule. Specifically, SMM matrix columns are sorted by their variance, and the
\code{k} columns with lowest variance are defined as anchor positions.
}
\examples{
## Investigate anchor positions for the HLA molecules A02:01 and B27:05
anchorPositions( "HLA-A-02:01" )
anchorPositions( "HLA-A-02:01", k=3 )
anchorPositions( "HLA-B-27:05" )
}

