% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-mhc.R
\name{plotBindingMotif}
\alias{plotBindingMotif}
\title{Plot an MHC Binding Motif}
\usage{
plotBindingMotif(mhc = "HLA-A-02:01", l = 9, motif.matrix = NULL,
  width = 0.5, space = 1, col = rainbow(20), main = paste(mhc, ", ", l,
  "-mers", sep = ""), ...)
}
\arguments{
\item{mhc}{name of the MHC molecule.}

\item{l}{length of the peptide.}

\item{motif.matrix}{a matrix. If this is not \code{NULL}, then the parameters
\code{mhc} and \code{l} are ignored and the supplied matrix is used instead. This
makes it possible to use this function for plotting sequence logos unrelated to
MHC-peptide binding. The supplied matrix must have row names to indicate the
letters out of which the sequence logo is made.}

\item{width}{bar width for profile per position.}

\item{space}{amount of space left before each bar, as a fraction of the
bar width.}

\item{col}{vector of colors to use for the amino acids (given in alphabetical order
in single-letter coding).}

\item{main}{title for the plot.}

\item{...}{further options to be passed on to \code{\link[graphics]{barplot}}}
}
\description{
Visualizes the set of peptides which are predicted to bind to a certain MHC molecule.
}
\examples{
## Compare binding motifs of HLA-A02 at all supported peptide lengths
par( mfrow=c(2,2) )
for( l in 8:11 ){
   plotBindingMotif( "HLA-A-02:01", l )
}
}

