% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-mhc.R
\name{smm}
\alias{smm}
\title{Peptide-MHC Binding Prediction}
\usage{
smm(x = c("SLYNTVATL", "SYFPEITHI"), mhc = "HLA-A-02:01",
  output.IC50 = TRUE)
}
\arguments{
\item{x}{vector of strings containing the peptides for which to predict MHC
binding.}

\item{mhc}{string or vector of strings identifying the MHC molecules. See
\code{supportedMHCs} for allowed values. If a vector is given, it must be of the same
length as \code{x}.}

\item{output.IC50}{whether to output the IC50 value itself (default) or its
base10-logarithm.}
}
\description{
Predicts peptide-MHC binding using the stabilized matrix method (SMM)
algorithm with a specifically constructed amino acid substitution matrix.
See \code{citation(package='EpitopePrediction')} for the reference.
}
\examples{
## Predict IC50 binding values for two famous peptides
smm( c("SLYNTVATL","SYFPEITHI"), "HLA-A-02:01" )
}

