% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-mhc.R
\name{smmMatrix}
\alias{smmMatrix}
\title{Get Prediction Matrix}
\usage{
smmMatrix(mhc = "HLA-A-02:01", l = 9)
}
\arguments{
\item{mhc}{name of the MHC molecule.}

\item{l}{the peptide length.}
}
\value{
A list with components \code{M} (the matrix) and \code{c} (a number to be
added to each prediction made with this matrix).
}
\description{
Loads the prediction matrix for a given MHC molecule at the given peptide length.
Matrices are loaded from the package folder upon first load and are then cached in
memory for further uses.
The SMM method predicts binding by summing the matrix entries for each amino acid
per position and then adding and MHC-dependant number. The resulting number is the
log IC50 value (IC50=half maximal inhibitory concentration). Hence, low numbers mean
strong binding and high numbers mean weak binding.
}
\examples{
## load prediction matrix for HLA-A-02:01 9-mers
M <- smmMatrix( "HLA-A-02:01", 9 )$M
## How well does Leucine bind at each position?
M["L",]
## Which amino acid is most preferred at the C-terminal position?
names( which.min( M[,9] ) )

## Naming of the MHC molecule is permissive
for( mhc in c("HLA-A0201","HLA-A-0201","HLA-A*0201","HLA-A*02:01") ){
   names( which.min( smmMatrix( "HLA-A-02:01", 9 )$M[,9] ) )
}

}

