% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-mhc.R
\name{supportedMHCs}
\alias{supportedMHCs}
\title{Supported MHC Molecules and Peptide Lengths}
\usage{
supportedMHCs(l = NULL)
}
\arguments{
\item{l}{integer giving the desired peptide length. If not given (default), then all
possible combinations of peptide lengths and MHC molecules are returned.}
}
\value{
A data frame containing the combinations of supported MHCs and peptide lengths.
}
\description{
Supported MHC Molecules and Peptide Lengths
}
\examples{
## Which MHC molecules are supported?
levels( supportedMHCs()$mhc )
## Which peptide lengths are supported for HLA-A02:01?
with( supportedMHCs(), l[mhc=="HLA-A-02:01"] )

}

